leadscore-api-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
sqlalchemy==2.0.30
pg8000==1.31.2

# Cloud SQL connector (requires Python 3.8–3.11)
cloud-sql-python-connector==1.9.0

# Google AI + logging
google-genai==0.3.0
python-json-logger==2.0.7
google-cloud-logging==3.9.0
google-cloud-error-reporting==1.9.1
